# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "析卷 AI",
...
    "version": "0.3.0"
  },
  "paths": {
    "/file/list": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "分页查询文件列表",
        "description": "按 `create_time DESC` 分页。`status` 与 `type_id` 均为**精确匹配**，空字符串表示不过滤。\n\n**`status` 合法值**（即 `progress` 字段）：`parsing` / `tableing` / `chunking` / `embedding` / `extracting` / `analyzing` / `complete`，以及对应的 `*_failed` 失败态。\n\n返回 `data=FileListResponse{items:[FileListItem], total, page, page_size, total_pages}`；`FileListItem = {file_id, file_name, file_size, progress, type_id, error, create_time}`。",
        "operationId": "list_files_file_list_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从 1 开始（默认 1）。"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页条数（默认 20）。"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            },
            "description": "按 `progress` 精确过滤：`parsing` / `tableing` / `chunking` / `embedding` / `extracting` / `analyzing` / `complete` 及对应 `*_failed`；空串不过滤。"
          },
          {
            "name": "type_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Type Id"
            },
            "description": "按文档类型精确过滤；空串返回全部类型。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/file/batch": {
      "delete": {
        "tags": [
          "file"
        ],
        "summary": "批量删除文件",
        "description": "删除请求体中所有 `file_id` 关联的 MySQL 记录、Milvus 向量、`uploads/{file_id}.pdf`。\n\n不存在的 `file_id` 不报错，会出现在响应 `failed_ids` 中。Milvus / PDF 清理走后台任务，失败仅记 warning，不影响 MySQL 删除。\n\n返回 `data=BatchDeleteResponse{deleted_count, failed_ids}`。",
        "operationId": "batch_delete_files_file_batch_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/file/parse": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "提交文件解析",
        "description": "上传文件并启动 6 阶段处理管线（parsing → tableing → chunking → embedding → extracting → analyzing → complete）。`Content-Type` 必须为 `multipart/form-data`。\n\n**`mode` 行为差异**\n- `async`（默认）：后台任务，立即返回 `data={file_id}`，`message=\"文件已提交处理（异步）\"`\n- `sync`：阻塞直到全部完成或失败，再返回 `data={file_id}`\n- `stream`：返回 `text/event-stream`，逐阶段推送事件（事件序列见 `API_DOCUMENTATION.md` 第 9.1 节）\n- 其它任意值按 `sync` 处理（代码 fallback 分支）\n\n**`callback_url`** 在 `async` 与 `sync` 模式下都会被使用（每阶段开始、每条 `field_done`/`rule_done`、每阶段 `stage_done` 都 POST 通知；超时 2.5s，失败仅 warning，不阻断主流程）。`stream` 模式忽略 `callback_url`，事件改走 SSE。\n\n**`file_id` 生成**：`SHA256[:32]((type_id, file_name, time.time_ns(), token_hex(8)))` —— 每次上传都是新 ID，**不做去重**。失败重试请用 `POST /file/{file_id}/retry/{stage}`，**不要**重新上传（会产生新记录）。\n\n原始 PDF 字节会同步写到 `uploads/{file_id}.pdf`（供 VL 抽取使用），写盘失败不阻断主管线。\n\n**文件大小限制**：受 `mineru.max_file_size` 控制（默认 100MB / 104857600 字节），超限返回 `ResponseWrapper{code:400, message:\"文件大小超过限制 (100MB)\"}`（不抛 HTTP 错误）。",
        "operationId": "parse_file_file_parse_post",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "async",
              "title": "Mode",
              "enum": [
                "async",
                "sync",
                "stream"
              ]
            },
            "description": "处理模式：`async`（默认，后台任务立即返回）/ `sync`（阻塞至完成）/ `stream`（SSE 流）。其它值按 `sync` 处理。"
          },
          {
            "name": "type_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Type Id"
            },
            "description": "归属文档类型，默认 `default`；决定使用哪套字段/规则配置。"
          },
          {
            "name": "callback_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Callback Url"
            },
            "description": "可选回调地址；管线每阶段开始 / `field_done` / `rule_done` / `stage_done` 都会向此 URL POST（超时 2.5s，失败仅 warning）。仅 `async` / `sync` 模式生效，`stream` 模式忽略。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_file_file_parse_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/file/{file_id}/status": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "查询文件处理进度",
        "description": "返回当前 `progress` 与最近的 `error`，含 `create_time` / `updated_at` 两个时间戳。\n\n`progress` 取值：`parsing` / `tableing` / `chunking` / `embedding` / `extracting` / `analyzing` / `complete`，以及对应的 `*_failed` 失败态。\n\n如需每阶段的开始/结束时间戳（计算阶段耗时），请走 `GET /file/{file_id}/detail`。\n\n**状态码**：200 / 404（文件不存在）。",
        "operationId": "get_file_status_file__file_id__status_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            },
            "description": "目标文件 ID（`POST /file/parse` 返回的 32 位 SHA256 摘要）。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/file/{file_id}": {
      "delete": {
        "tags": [
          "file"
        ],
        "summary": "删除文件",
        "description": "级联删除：`files` / `file_content` / `file_table` / `file_chunk` / `extraction_result` / `analysis_result`（MySQL 立即提交），Milvus 向量与 `uploads/{file_id}.pdf` 走后台清理（失败仅 warning），接口立即返回。\n\n**状态码**：200 / 404（文件不存在）。",
        "operationId": "delete_file_file__file_id__delete",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            },
            "description": "目标文件 ID（`POST /file/parse` 返回的 32 位 SHA256 摘要）。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/file/{file_id}/retry/extracting": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "快捷重试：字段提取",
        "description": "语义等价于 `POST /file/{file_id}/retry/{stage}` 中 `stage=extracting`，内部即转发调用。\n\n支持参数：`mode`（async/sync/stream）、`callback_url`（async/sync 生效）。",
        "operationId": "retry_extracting_file__file_id__retry_extracting_post",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            },
            "description": "目标文件 ID（`POST /file/parse` 返回的 32 位 SHA256 摘要）。"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "async",
              "title": "Mode",
              "enum": [
                "async",
                "sync",
                "stream"
              ]
            },
            "description": "处理模式：`async`（默认）/ `sync` / `stream`。"
          },
          {
            "name": "callback_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Callback Url"
            },
            "description": "可选回调地址；管线每阶段开始 / `field_done` / `rule_done` / `stage_done` 都会向此 URL POST（超时 2.5s，失败仅 warning）。仅 `async` / `sync` 模式生效，`stream` 模式忽略。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/file/{file_id}/analysis": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "逻辑分析结果",
        "description": "返回 `analysis_result` 全表行：`data=[{file_id, rule_id, result_value, input_values, reason}]`（含 `input_values` 字典，但不含 `source_refs`）。",
        "operationId": "get_analysis_results_file__file_id__analysis_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            },
            "description": "目标文件 ID（`POST /file/parse` 返回的 32 位 SHA256 摘要）。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchDeleteRequest": {
        "properties": {
          "file_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "File Ids",
            "description": "要删除的文件 ID 列表（不存在的会进入响应 `failed_ids`）。"
          }
        },
        "type": "object",
        "required": [
          "file_ids"
        ],
        "title": "BatchDeleteRequest",
        "description": "批量删除文件的请求体。",
        "examples": [
          {
            "file_ids": [
              "a1b2c3...",
              "d4e5f6..."
            ]
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseWrapper": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200,
            "description": "业务状态码，成功 200；业务校验失败（如文件超限）也可能在 200 响应里返回非 200 的 `code`。"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success",
            "description": "人类可读的结果说明。"
          },
          "data": {
            "title": "Data",
            "description": "业务负载，类型随接口而定（对象 / 数组 / null）。"
          }
        },
        "type": "object",
        "title": "ResponseWrapper",
        "description": "统一响应包装。业务负载在 `data` 中，其具体形态见各接口描述。"
      },
...
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:5019",
      "description": "本地开发"
    }
  ]
}